# product/src/main/resources/swagger/api.yml
openapi: '3.0.3'

info:
  title: 'product'
  version: 0.0.1

servers:
  - url: http://localhost:10002
    description: Product Server

paths:

  /api/products:
    post:
      description: 개별 상품 등록
      operationId: createProduct
      tags:
        - Product
      requestBody:
        content:
          application/vnd.vroong.private.v1+json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          $ref: '#/components/responses/Created'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

    get:
      description: 상품 목록 조회
      security: []
      operationId: listProducts
      tags:
        - Product
      parameters:
        - $ref: '#/components/parameters/SearchWordParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/PageParam'
      responses:
        '200':
          description: OK
          content:
            application/vnd.vroong.private.v1+json:
              schema:
                $ref: '#/components/schemas/ProductList'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'

  /api/products/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductId'

    get:
      description: 개별 상품 조회
      security: []
      operationId: getProduct
      tags:
        - Product
      responses:
        '200':
          description: OK
          content:
            application/vnd.vroong.private.v1+json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'


    patch:
      description: 개별 상품 수정
      operationId: updateProduct
      tags:
        - Product
      requestBody:
        content:
          application/vnd.vroong.private.v1+json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: OK
          content:
            application/vnd.vroong.private.v1+json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      description: 개별 상품 삭제
      operationId: deleteProduct
      tags:
        - Product
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

components:
  schemas:

    #----------------------------------------
    # API fields
    #----------------------------------------
    Size:
      type: object
      properties:
        width:
          $ref: '#/components/schemas/IntegerField'
        height:
          $ref: '#/components/schemas/IntegerField'
        depth:
          $ref: '#/components/schemas/IntegerField'

    #----------------------------------------
    # Common Model
    #----------------------------------------
    LongId:
      type: integer
      format: int64
      example: 1

    IntegerField:
      type: integer
      format: int32
      example: 1

    DateTime:
      type: string
      format: date-time
      example: "2018-12-01T10:00:00Z"

    StringField:
      type: string
      example: "foo"

    UUID:
      type: string
      format: uuid
      example: c9106593-2d6c-40eb-8504-66b0ef41ad2d

    Money:
      type: number
      example: 1000

    LongArray:
      type: object
      properties:
        ids:
          type: array
          items:
            $ref: '#/components/schemas/LongId'

    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
          example: product/constraint-violation
        title:
          type: string
          example: 제출한 요청이 유효하지 않습니다
        detail:
          type: string
        fieldErrors:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/FieldError'

    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          $ref: '#/components/schemas/ErrorCode'

    ErrorCode:
      type: string
      example:
        - NotNull
        - Size

    Page:
      type: object
      properties:
        size:
          type: integer
          format: int32
          default: 20
          example: 20
        totalElements:
          type: integer
          format: int64
          example: 100
        totalPages:
          type: integer
          format: int32
          example: 5
        number:
          type: integer
          format: int32
          default: 1
          example: 1

    #----------------------------------------
    # API Model
    #----------------------------------------

    Example:
      type: object
      properties:
        exampleId:
          $ref: '#/components/schemas/LongId'
        title:
          $ref: '#/components/schemas/StringField'
        createdAt:
          $ref: '#/components/schemas/DateTime'
        updatedAt:
          $ref: '#/components/schemas/DateTime'
        createdBy:
          $ref: '#/components/schemas/StringField'
        updatedBy:
          $ref: '#/components/schemas/StringField'

    Product:
      type: object
      properties:
        productId:
          $ref: '#/components/schemas/LongId'
        name:
          $ref: '#/components/schemas/StringField'
        description:
          $ref: '#/components/schemas/StringField'
        inventory:
          $ref: '#/components/schemas/IntegerField'
        price:
          $ref: '#/components/schemas/Money'
        supplier:
          $ref: '#/components/schemas/StringField'
        size:
          $ref: '#/components/schemas/Size'
        location:
          $ref: '#/components/schemas/StringField'
        created_at:
          $ref: '#/components/schemas/DateTime'
        updated_at:
          $ref: '#/components/schemas/DateTime'

    ProductList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          $ref: '#/components/schemas/Page'

  parameters:

    PageParam:
      in: query
      name: page
      schema:
        type: integer
        format: int32
        default: 1

    SizeParam:
      in: query
      name: size
      schema:
        type: integer
        format: int32
        default: 10

    SearchWordParam:
      in: query
      name: q
      schema:
        $ref: '#/components/schemas/StringField'

    ProductId:
      in: path
      name: productId
      schema:
        $ref: '#/components/schemas/LongId'
      required: true

  responses:

    Created:
      description: Created

    Accepted:
      description: Accepted

    NoContent:
      description: No Content

    EmptyContent:
      description: No Content
      content:
        application/vnd.vroong.private.v1+json: { }

    BadRequest:
      description: Bad Request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

    Unauthorized:
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

    Forbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

    NotFound:
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

    Conflict:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

    ServerError:
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'

  securitySchemes:
    'jhipster-uaa':
      type: oauth2
      flows:
        password:
          tokenUrl: http://localhost:9999/uaa/oauth/token
          refreshUrl: http://localhost:9999/uaa/oauth/token
          scopes: {}
        clientCredentials:
          tokenUrl: http://localhost:9999/uaa/oauth/token
          scopes: {}

security:
  - jhipster-uaa: []